// src/app/models/Todo.model.ts
export class Todo {

    id: number;
    title: string;
    completed: boolean = false;
    userId?: number;

    // constructor(id: number, title: string, userId: number) {
    //     this.id = id;
    //     this.title = title;
    //     this.userId= userId;
    // }

    // isComplete(completed: boolean) {
    //     this.completed = completed;
    // }
}